Split CLIFormatter implementations into different files.
